requirements.txt
asgiref==3.8.1
cffi==1.17.1
coverage==7.9.1
cryptography==45.0.5
Django==5.2.3
django-admin-honeypot-updated-2021==1.2.0
django-cors-headers==4.7.0
django-environ==0.12.0
django-ipware==7.0.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
gunicorn==23.0.0
mysqlclient==2.2.7
packaging==25.0
pillow==11.3.0
pycparser==2.22
PyJWT==2.9.0
python-dotenv==1.1.1
python-ipware==3.0.0
pytz==2025.2
sqlparse==0.5.3
tzdata==2025.2
whitenoise==6.9.0
